Enemy Mage Mesh Updated
The enemy mage mesh has been updated to a new mesh with no arrows. The
only downside is that the old texture in the area where the arrows used
to be is a little weird. If someone can fix that be my guest. I also
retarget the animations to the new mesh so once changing over the new
mesh is complete you can just delete all the old files.