Small Fixes and Tweaks
The player will no longer be able to see through an object when they get
too close to it. I also tweaked the max distance of telekinesis to be a
little bit further, before it seemed too confined. Last I changed the
scroll value for bringing an object closer or pusing it away to half its
original value. This is due to the fact that a key button will fire more
and much faster than the mouse wheel.