Level Design
-Updated Level 1. Nearly finished, need to add touches like player
dialogue for tutorials and other things. First puzzle complete, super
simple but should be effective.
-LevelExit blueprint added. Acts as the gateway to the next level. To
use, place in level, look in the inspector...editor thing and find the
input for the level name. When the player presses E near the exit, the
next level will open.
-PlayerSpeechText added. It is a UI element that will show at the bottom
of the screen to represent player dialogue. It will display a sequence
of text items then fade out.
-PlayerTextTrigger is a trigger box that will trigger player dialogue.
To use, place the box in the level and scale as you wish. In the editor,
see the array of strings that it accepts. Fill the array with dialogue
elements for each bit of speech the player will give. On collision with
the trigger, the dialogue will play.
-Edited the size and position of the DropDownText.