Fixed Some Things, Added Some Things
I fixed the issue with getting a null error when using magic on a
non-MovableObjectBP object (aka floatingObject). Coincidentally that
fixes the issues with any object that is movable and not a
movableObjectBP. Therefore any object can now be movable and not be a
movableObjectBP i.e. the object can move but magic will not affect it
(push I am not sure about though). I also added an option to add
rotation to the floatingObjectBP