LargeTextTrigger and Game Instance
- Added LargeTextTrigger blueprint. This blueprint is a trigger box
that, upon collision with the player, creates a UI element like the one
that plays when you start the level. This can also be used to add to the
counter for puzzles complete. To use, place in the world and see the
public self explanatory variables.
- Created a game instance to store progress for the player and to use as
a marker to block certain spells.